clean up

+++ assert.go
@@ -82,7 +82,7 @@ func AssertEqual(t *testing.T, expected any, actual any) {
 	t.Helper()
 
 	if !isEqual(expected, actual) {
-		logf(t, decorateDiff(expected, actual))
+		log(t, decorateDiff(expected, actual))
 	}
 }
